feat: deploy Qwen3-30B-A3B-FP8, fix startup warnings, add kernel tuning scripts
Deployment
- Switch to Qwen/Qwen3-30B-A3B-Instruct-2507-FP8 on GPUs 1+2 (TP=2, RTX 4090)
- Ports moved to 8010 (vLLM) / 8020 (gateway) to avoid conflicts
- Add VLLM_URL to .env so the gateway resolves the upstream correctly

Bug fixes
- VLLM_CACHE_ROOT redirected to /tmp2 to prevent disk-quota crash on home partition
- CUDA_DEVICE_ORDER=PCI_BUS_ID added for stable indexing on mixed 4090+3090 fleet
- Remove deprecated TRANSFORMERS_CACHE from launch_vllm.sh (use HF_HOME only)
- pyproject.toml: migrate [tool.uv.dev-dependencies] → [dependency-groups]

vLLM launch hardening
- --disable-custom-all-reduce: silence P2P-unavailable warning on consumer GPUs
- --generation-config vllm: prevent HF model card from silently overriding sampling params
- VLLM_CACHE_ROOT export + mkdir -p in launch_vllm.sh

Triton kernel configs for RTX 4090
- scripts/write_kernel_configs.py: instant heuristic JSON configs (no GPU required)
  suppresses "sub-optimal" FP8/MoE warnings; run once after install
- scripts/tune_kernels.py: benchmark-tuned configs via GPU timing (5-10 min);
  uses tight 24-config search space on 5 representative batch sizes, extrapolates rest

Makefile
- make status and make loadtest now read VLLM_PORT/GATEWAY_PORT from .env
  instead of hardcoding 8000/8080

Docs
- README: updated hardware table, ports, model, vLLM version, uv launch commands
- Document VLLM_CACHE_ROOT requirement, kernel config workflow, known RTX 4090 limits
  (CutlassBlockScaledGroupedGemm needs Hopper TMA; FlashInfer has no torch 2.7 wheel yet)

+++ rust/gpu-exporter/src/main.rs
@@ -165,7 +165,7 @@ async fn collect_once(metrics: &GpuMetrics) {
         macro_rules! set {
             ($metric:expr, $idx:expr) => {
                 if let Ok(v) = safe_float(parts[$idx]) {
-                    $metric.get_or_create(&labels).set(v);
+                    $metric.get_or_create(&labels).set(v as i64);
                 }
             };
         }
